build: Create @profiscience/knockout-contrib metapackage
Having another go at metapackages, but sticking manual this time. It's a little tedious, but in
reality new packages aren't added all that often and in the end the flexibility gained is well worth
it. By default, the metapackage will (should) install the latest version of each package, unless
there is a version otherwise specified either in the user's package.json, or any cross-dependencies.
We'll see if it works how I hope it does. A boilerplate file has been added.