build: Build declarations and declaration maps for consumers
Pointing TS consumers at the source requires them to configure their tsconfig.json for this project,
which isn't ideal. Declarations should avoid this by preventing consumer's compilers from seeing the
actual implementation details. In particular, this should prevent consumers from having to mess with
lib or set allowSyntheticDefaultImports (because of path-to-regexp in the router package).