Revert "Emit .d.ts files for consumers (#93)"
This reverts commit 9bb8e3c5f22d369e3db75a0a9513ce6007237550.

It's just not going to work right now. For some reason, you can not
emit declarations with the mixin pattern if that mixin uses a protected
method, which many do in our case since `.fetch()` is marked protected
in the base data model.

Also, since the declarations were emitted adjacent to the source, _only_
the entry (from package.json) was pointing to the declaration. From there,
the entry's declaration (index.d.ts) import paths (import './someotherfile')
would resolve to the original source, defeating the purpose.

For now, the best thing is to use the source for the type declarations,
and document the requirements in the tsconfig for consumers.