fix(router.plugins.component): Hoist viewModel.dispose() method to beforeDispose lifecycle phase
Because the execution of middleware overlaps when navigating, using the dispose method can lead to
unexpected behavior. For example, when navigating from A to B, if both A and B use the same package
w/ side effects that require a clean slate, component B's "beforeRender" will execute before the
"dispose" method on component A's viewModel.