fix(router): Dispose initiated child middleware during redirection
When using nested routing and middleware which queues async results, it's
likely that child middleware gets executed _before_ an async function in
one of the parent routes calls ctx.redirect. Normally, the after render
middleware is called by knockout when the component renders, naturally
calling them down the cascade. But, since there _is no_ render during
redirection, it must be called explicitly (to make sure any necessary cleanup
occurs). In this case, it's necessary to explicitly call it for child
contexts as well.